More changes and refactoring of the kubernetes scheduler. Json info is initially set (untested)
