+++ heron/schedulers/src/java/com/twitter/heron/scheduler/kubernetes/KubernetesConstants.java
@@ -49,14 +49,14 @@ public final class KubernetesConstants {
       "master", "tmaster-controller", "tmaster-stats", "shell", "metricsmgr", "scheduler",
       "metrics-cache-master", "metrics-cache-stats"};
 
-  public static final String MASTER_PORT = "$PORT0";
-  public static final String TMASTER_CONTROLLER_PORT = "$PORT1";
-  public static final String TMASTER_STATS_PORT = "$PORT2";
-  public static final String SHELL_PORT = "$PORT3";
-  public static final String METRICSMGR_PORT = "$PORT4";
-  public static final String SCHEDULER_PORT = "$PORT5";
-  public static final String METRICS_CACHE_MASTER_PORT = "$PORT6";
-  public static final String METRICS_CACHE_STATS_PORT = "$PORT7";
+  public static final String MASTER_PORT = "6001";
+  public static final String TMASTER_CONTROLLER_PORT = "6002";
+  public static final String TMASTER_STATS_PORT = "6003";
+  public static final String SHELL_PORT = "6004";
+  public static final String METRICSMGR_PORT = "6005";
+  public static final String SCHEDULER_PORT = "6006";
+  public static final String METRICS_CACHE_MASTER_PORT = "6007";
+  public static final String METRICS_CACHE_STATS_PORT = "6008";
 
   public static final String[] PORT_LIST = new String[]{
       MASTER_PORT, TMASTER_CONTROLLER_PORT, TMASTER_STATS_PORT,
