+++ heron/schedulers/src/java/com/twitter/heron/scheduler/kubernetes/KubernetesLauncher.java
@@ -14,8 +14,13 @@
 
 package com.twitter.heron.scheduler.kubernetes;
 
+import java.nio.file.Paths;
+import java.util.logging.Logger;
+
 import com.twitter.heron.scheduler.utils.LauncherUtils;
+import com.twitter.heron.scheduler.utils.SchedulerUtils;
 import com.twitter.heron.spi.common.Config;
+import com.twitter.heron.spi.common.Context;
 import com.twitter.heron.spi.packing.PackingPlan;
 import com.twitter.heron.spi.scheduler.ILauncher;
 import com.twitter.heron.spi.scheduler.IScheduler;
@@ -24,13 +29,20 @@ import com.twitter.heron.spi.scheduler.IScheduler;
  * Submit topology to Marathon.
  */
 public class KubernetesLauncher implements ILauncher {
+  private static final Logger LOG = Logger.getLogger(KubernetesLauncher.class.getName());
+
   private Config config;
   private Config runtime;
+
+  private String schedulerWorkingDirectory;
 
   @Override
   public void initialize(Config aConfig, Config aRuntime) {
     this.config = aConfig;
     this.runtime = aRuntime;
+
+    // get the scheduler working directory
+    this.schedulerWorkingDirectory = KubernetesContext.getSchedulerWorkingDirectory(config);
   }
 
   @Override
@@ -40,13 +52,31 @@ public class KubernetesLauncher implements ILauncher {
 
   @Override
   public boolean launch(PackingPlan packing) {
+
+    // setup the working directory -- mainly just need to go and grab the topology jar
+    if (!setupWorkingDirectory()) {
+      LOG.severe("Failed to setup working directory");
+      return false;
+    }
+
     LauncherUtils launcherUtils = LauncherUtils.getInstance();
     Config ytruntime = launcherUtils.createConfigWithPackingDetails(runtime, packing);
     return launcherUtils.onScheduleAsLibrary(config, ytruntime, getScheduler(), packing);
   }
 
-  // Get MarathonScheduler
+  // Get KubernetesScheduler
   protected IScheduler getScheduler() {
     return new KubernetesScheduler();
+  }
+
+  // Setup the working directory -- pull down the topology jar
+  protected boolean setupWorkingDirectory() {
+    String topologyPackageURL = String.format("file://%s", Context.topologyPackageFile(config));
+    String topologyPackageDestination = Paths.get(
+        schedulerWorkingDirectory, "topology.tar.gz").toString();
+
+    return SchedulerUtils.curlAndExtractPackage(
+        schedulerWorkingDirectory, topologyPackageURL,
+        topologyPackageDestination, true, Context.verbose(config));
   }
 }
